split page and widgets
