+++ lib/models/finance.dart
@@ -1,0 +1,8 @@
+class Finance {
+  double today;
+  double yesterday;
+
+  double get diff => today - yesterday;
+
+  double get diffRate => ((today / yesterday) - 1) * 100;
+}
