+++ lib/styles/color.dart
@@ -1,0 +1,3 @@
+import 'package:flutter/material.dart';
+
+const Color textColor = Color(0xFFFAFAFA);
